pyproject.toml
[project]
name = "document-llm-rag-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.2.1",
    "fastapi>=0.115.12",
    "jinja2>=3.1.6",
    "minio>=7.2.15",
    "openai>=1.84.0",
    "pydantic-ai>=0.2.12",
    "pydantic-settings>=2.9.1",
    "pypdf>=5.6.0",
    "python-dotenv>=1.1.0",
    "qdrant-client>=1.14.2",
]

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "pytest>=8.3.5",
    "pytest-asyncio>=1.0.0",
    "testcontainers>=4.10.0",
]

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "DEBUG"
log_cli_format = "%(asctime)s [%(levelname)s] %(filename)s:%(lineno)s %(message)s"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
testpaths = ["tests"]

[tool.pytest-env]
MODE = "dev"

[tool.pytest-asyncio]
asyncio_mode = "auto"
default_test_loop_scope = "session"
default_fixture_loop_scope = "session"

[tool.ruff]
line-length = 88
indent-width = 4

target-version = "py313"

[tool.ruff.lint]
select = ["ALL"]
ignore = ["D", "PGH003"]
fixable = ["ALL"]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = false
docstring-code-line-length = "dynamic"

[tool.ruff.lint.per-file-ignores]
"**/{tests,docs,tools}/*" = ["S101", "SLF001", "ANN", "ARG", "PTH", "PERF", "PLR0913", "PLR2004"]
